+++ Pandora/Pattern/Operation.hs
@@ -2,6 +2,10 @@ module Pandora.Pattern.Operation (module Exports) where
 
 import Pandora.Pattern.Operation.Zero as Exports
 import Pandora.Pattern.Operation.One as Exports
+import Pandora.Pattern.Operation.Unit as Exports
 import Pandora.Pattern.Operation.Sum as Exports
 import Pandora.Pattern.Operation.Product as Exports
 import Pandora.Pattern.Operation.Exponential as Exports
+
+type instance Unit (:*:) = One
+type instance Unit (:+:) = Zero