+++ Pandora/Pattern/Operation/Unit.hs
@@ -1,0 +1,3 @@
+module Pandora.Pattern.Operation.Unit where
+
+type family Unit (p :: * -> * -> *) = r | r -> p